% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validate_library_frame}
\alias{validate_library_frame}
\title{Validate Library names/locations in a Dataframe}
\usage{
validate_library_frame(counts, design, countsField = "colnames",
  designField = "lib", metrics = metrics, summary = summary,
  metricsField = "libid", summaryField = "libid")
}
\arguments{
\item{counts}{counts dataframe}

\item{design}{design dataframe}

\item{countsField}{counts library field - usually column names}

\item{designField}{design library field - either the row names, or a column in the dataframe}

\item{metrics}{metrics dataframe (optional)}

\item{summary}{summary dataframe (optional)}

\item{metricsField}{metrics library field - either the row names, or a column in the dataframe}

\item{summaryField}{metrics library field - either the row names, or a column in the dataframe}
}
\value{
boolean
}
\description{
validate standard RNA-seq data: counts, design, metrics, and summary files.
}
\author{
Scott R Presnell, \email{SPresnell@benaroyaresearch.org}
}
