% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validate_library_names}
\alias{validate_library_names}
\title{Validate Library names/locations by name}
\usage{
validate_library_names(countsNames, designNames, metricsNames = NULL,
  summaryNames = NULL)
}
\arguments{
\item{countsNames}{counts name list}

\item{designNames}{design name list}

\item{metricsNames}{metrics name list}

\item{summaryNames}{summary name list}
}
\value{
boolean
}
\description{
validate standard RNA-seq data: counts, design, metrics, and summary files.
}
\author{
Scott R Presnell, \email{SPresnell@benaroyaresearch.org}
}
