% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validateLibraryFrame}
\alias{validateLibraryFrame}
\title{Validate Library names/locations in a Dataframe}
\usage{
validateLibraryFrame(counts, design, countsField = "colnames",
  designField = "libraryId", metrics = NULL, summary = NULL,
  metricsField = "libraryId", summaryField = "libraryId")
}
\arguments{
\item{counts}{counts dataframe}

\item{design}{design dataframe}

\item{countsField}{counts library field - usually column names}

\item{designField}{design library field - either the row names, or a column in the dataframe}

\item{metrics}{metrics dataframe (optional)}

\item{summary}{summary dataframe (optional)}

\item{metricsField}{metrics library field - either the row names, or a column in the dataframe}

\item{summaryField}{metrics library field - either the row names, or a column in the dataframe}
}
\value{
boolean
}
\description{
validate standard RNA-seq data: counts, design, metrics, and summary files.
}
\author{
Scott R Presnell, \email{SPresnell@benaroyaresearch.org}
}
