% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CamelNames.R
\name{CamelFrame}
\alias{CamelFrame}
\title{Convert data frame names to camelCase}
\usage{
CamelFrame(dframe)
}
\arguments{
\item{dframe}{dataframe whose names to convert to camelCase}
}
\value{
data frame with new names
}
\description{
Convert column names of a data frame to camelCase by taking them all to lower case, and looking for word boundaries and makeing those upper case.
Then trim off the white space.
}
\author{
Scott R Presnell, \email{SPresnell@benaroyaresearch.org}
}
